Add getInheritableValue method for looking up the value of a data element that can be inherited from the parent node in the hierarchy.
Add performRenames() method to allow existing data elements to be
renamed. Resave the data file if any such renames occur. (Thus, renaming
should typically be used only for end-user-editable data, NOT for changing
the name of a data element in a datafile you inherit from.)

When saving a datafile, include a comment stating what task it belongs to.

Extra tweaks to avoid realizing deferred data when removing an element or
saving a datafile.


git-svn-id: https://svn.code.sf.net/p/processdash/code/trunk@239 560cad82-bde2-470f-af40-e4e1f8d6548a