New improved mechanism for searching for process templates. Now it will try the following search path:
(1) Look first in the directory specified by the user setting
    "templates.directory" (should end in "Templates").
(2) Next, look in any JAR files contained in that directory.
(3) Next, look in any JAR files contained in the parent of that directory.
(4) If there is a Templates directory alongside the pspdash.jar file,
    look there, first for the file, then for JARs containing the file.
(5) If there are any JAR files next to the pspdash.jar file, look in them.
(6) Finally, look in the classpath (which includes the pspdash.jar file).


git-svn-id: https://svn.code.sf.net/p/processdash/code/trunk@278 560cad82-bde2-470f-af40-e4e1f8d6548a