Internet Explorer expects that the RWStatus of a field is static.  This means that it reads the RWStatus of a field once (typically in response to the datasetComplete or datasetChanged event), then never checks again.  When the value of a field changes, IE requeries to get the new value, but does not requery to see if the editability has changed.
In v1.2 of the dashboard, this was not a problem because data elements never
changed from editable to read-only or vice versa. But when the freezing
mechanism was introduced in v1.3, IE started having problems.  Take the
scenario where data was frozen when you opened up a browser window, then you
thawed the data with the window still open.  The INPUT fields would correctly
become read only (since that is handled by JavaScript).  But if you edited
the data, IE would complain that the underlying element in the data source
was read-only, and would discard your change.

The solution is that we tell the IE data binding logic that everything is
editable; then we handle editability ourselves, in both JavaScript and in the
DataInterpreter logic.


git-svn-id: https://svn.code.sf.net/p/processdash/code/trunk@922 560cad82-bde2-470f-af40-e4e1f8d6548a