Support custom size metrics for a team process.
Note: this implementation is slightly incompatible with earlier
functionality.  Previously, the WBS would store estimated size data for
a requirements document in the attribute "Reqts Pages".  Now, it will be
stored as "Req Pages".  This will cause the WBS to forget estimated
sizes for requirements documents in the WBS for historical projects.
Note that the actual project metrics (as viewed in the project plan
summary) will not be affected - just views of the WBS.

git-svn-id: file:///teamdash/trunk@313 9a68f182-09ab-e74b-aa6c-e8177106e7d0