Fix a usability problem with the "paste" operation.
The tree in the WBS editor is designed to make it easy for individuals to type
node names; it immediately goes into edit mode on a click.  However, this
doesn't play well with recent changes, which employ the system clipboard to
copy and paste lists of nodes.

Specifically, when a user copies a list of nodes to the clipboard (either from
an external source like Excel, or from within the WBS itself), these nodes can
be retrieved from the clipboard as String data.  Then, if the user has clicked
on a WBS node and is in "node name editing mode", typing Ctrl-V will paste a
huge string of garbage into the node name.

This change fixes that behavior.  If the text on the system clipboard looks
like a set of node names, the node name editor will refuse to accept it, and
will allow the WBS tree to handle the paste operation instead.

git-svn-id: file:///teamdash/trunk@402 9a68f182-09ab-e74b-aa6c-e8177106e7d0