Add autocomplete support to the "Labels" column in the WBS.
Also cleaned up the separation of concerns for UndoList processing in the DataJTable.  Previously, UndoList integration was handled in the table
cell editor.  Unfortunately, this was a constant source of bugs: when
new "custom edited columns" are introduced, it is common to forget to add
the UndoList integration.  In addition, the UndoList integration code was
copied and pasted into multiple locations.  Now, the UndoList integration
has been moved into DataJTable itself.

git-svn-id: file:///teamdash/trunk@559 9a68f182-09ab-e74b-aa6c-e8177106e7d0