Temporarily disable the "http.allowRemote" user setting.
Today, remote connections must be enabled to support the most common method
of joining individuals to a team project.  If a team leader unchecks this box
without understanding the implications, the joining process will fail.  Thus,
the presence of this checkbox could complicate the task of configuring the
Team Dashboard.

In the future, it will be helpful to create a new joining process that
does not require P2P connectivity.  Once that is in place, this checkbox
can be reenabled.

git-svn-id: https://svn.code.sf.net/p/processdash/code/trunk@3536 560cad82-bde2-470f-af40-e4e1f8d6548a