Enhance the Team Member List editor so it can lookup users from an external directory.  This is accomplished as a collaboration with the PDES, which actually performs the lookup.  When users are lookup up in this way, display a small icon next to their name in the team member list.
This change also introduces a small bit of forward compatibility for the team
member list: if we load the team member list from XML and we don't recognize
some of the attributes we find, keep them around and write them back out again.
This will help us to avoid clobbering data that was written by someone else
using a newer version of the WBS editor.

git-svn-id: file:///teamdash/trunk@1132 9a68f182-09ab-e74b-aa6c-e8177106e7d0