Some custom metrics collection frameworks contain parameter references to external files.  Those compile just fine when run in a build process against the filesystem.  However, when an end-user opens the resulting metrics collection framework in the Custom Metrics Collection Framework Editor, makes changes, and saves the result, the servlet-based compiler cannot locate the external files that were referenced by the custom framework parameter.
The end result is that when an end-user opens the TSP.zip framework, modifies
it, and saves the result, the resulting framework will encounter an error
whenever they attempt to view the rollup plan summary for a team project.
This is a significant problem.

This change introduces special handling for the servlet-based compilation
mode.  Specifically, when we are running in a servlet process, all "filename"
based custom parameters are silently ignored so the generation script can
fall back to the default behavior.

git-svn-id: file:///teamdash/trunk@1344 9a68f182-09ab-e74b-aa6c-e8177106e7d0