Introduce a new option for pie charts, to assign wedge colors "predictably."
Normally, JFreeChart assigns colors to pie wedges as needed.  If a wedge is of zero size, the color gets assigned to the next wedge.  However, there are cases where we are drawing several associated pie charts that all have the same category labels, and we would like a particular category to be assigned the same color always (regardless of whether any zero wedges preceded it).  The new flag makes this possible.

git-svn-id: https://svn.code.sf.net/p/processdash/code/trunk@4456 560cad82-bde2-470f-af40-e4e1f8d6548a