Allow MCF XML files to specify a process timestamp
The version number of an MCF add-on ZIP file contains a timestamp
indicating when the MCF was saved. For individuals using the MCF editor,
this is useful because it captures the time the process (phases, etc)
was last modified. In the past, this timestamp was also useful to
capture the time the MCF was regenerated during the build process. But
with the new lightweight packaging approach, MCF add-ons are no longer
regenerated during the build, and changes to this timestamp would just
trigger unnecessary upgrade warnings for individuals joining a team
project. This commit allows a fixed timestamp to be stored in the XML
file, recording the time when the process last changed.