Retain inferred MCF XML files in data backups
Data backup files include copies of the MCF XML process definitions that
are in use by the given dataset. When those XML files are used in the
place of an installed MCF add-on, and the user saves a new data backup,
ensure that the MCF XML files make it into the new backup too.