Bugfix: sync error triggered by MCF with long name
When the Team Dashboard performs a sync to WBS operation, it has
historically created a set of -template.xml files in the network
directory to describe the various team project workflows. When the name
of an MCF is unusually long, a quirk in template resolution could cause
one of these workflows to be incorrectly selected as the official
template for the "Indiv2Task" template. As a result, the sync operstion
would incorrectly instantiate the given workflow underneath every task
in the project. This change alters the name of the real Indiv2Task
element to avoid that possibility.