Display WBS change hist report in the script list
This change displays an entry in the team and personal dashboard script
lists for the WBS Change History Report. Selecting that item displays the
change history for the active project. Error handling has also been added
to display appropriate error messages when the necessary files cannot be
found or read.