Bugfix: user-completed nodes should not be sync-deletable
When performing a sync operation, some nodes are marked for deletion (if
they contain no actual data) or marked complete (otherwise). In the
past, a user-entered completion date was not considered to be "actual
data," so completed tasks could be deleted if they had no time logged.
This would lead to mass deletion of completed-with-zero-time tasks after
a project had been relaunched. This change fixes that bug.