Merge tester class can now save the merged results
The TeamProjectMergeTester class has historically been used to run a
merge and see if any exceptions occur (to fix bugs). As such, if no
errors were encountered, it would exit silently. This change adds the
feature to save the merged results into files on the hard drive. This
makes it possible to hand-merge project files for special ad-hoc needs.