Use the name of the current user for WBS save operations
When the WBS Editor saves data, it creates an entry in the change
history file with the name of the individual who saved it. In the past,
this value would be the name of a personal dashboard owner if the WBS
was opened from a personal dashboard. Now the value will always come
from a preference attached to the currently logged-on user.