In bridged mode, guard against local working dir disappearance
When running in bridged mode against a server-based dataset, files are
copied and cached in a directory on the local filesystem. If an unusual
event causes that directory to become unavailable, detect this and
trigger a "cannot save data" error. (It is possible in the past that
this scenario could trigger a massive deletion of files from the dataset
on the server.)