Do not assign the dataset ID eagerly on first startup
Beginning in version 2.1.7, the dashboard was eagerly writing a new
datasetID.dat file for each new dataset the first time that dataset was
opened. Although this seemed harmless, it increased the probability of a
severe problem scenario. Specifically, if a coach created and configured
an empty personal dashboard, then made copies of it for each team
member, then all individuals would share the same dataset ID. This would
lead to serious problems with warehouse-driven data calculations, which
rely on the dataset ID to be unique among individuals.

This change eliminates the eager assignment of dataset IDs, which are
now likely to be assigned only as an individual joins their first team
project. This change will make the scenario above much less likely to
occur. Team dashboards also do not receive dataset IDs until the first
project is created within.