Guard against null bytes written to data files
In the past, users have occasionally seen corrupt data files filled with
the null byte (ASCII 0). Suspicions in the past have pointed to
corruption from an external source (such as a virus scanner), but no
root cause has been found. On the chance that the null bytes are indeed
being written by Java (for example, triggered by an OutOfMemory or
similarly severe error), introduce an object that will detect this error
and prevent the null bytes from being written to disk. The current
change only checks the initial byte written to a data file, but even
that check should be sufficient to trap the occurrences of this problem
that have been observed in the past.