Preserve a minimum number of old backup files.
On startup, the dashboard will normally clean up backup files that are
more than 42 days old. Unfortunately, if a particular dataset has not
been viewed in the last 42 days, that means it will discard ALL old
backup files. This change keeps a minimum of 42 files, even if those
files happen to be very old.