When files contain null bytes, attempt to restore from backup
In the past, users have occasionally seen corrupt data files filled with
the null byte (ASCII 0). Suspicions in the past have pointed to
corruption from an external source (such as a virus scanner), but no
root cause has been found. On startup, if any such files are found, look
for a suitable replacement within a ZIP in the backup subdirectory. If
this repair is successful, display a warning to the user letting them
know that data may have been lost.