Use a Launch4J wrapper to open the Process Dashboard
Recent changes by Oracle have caused shortcuts to javaw to fail on some
versions of Windows. To work around this problem, we now create a
Launch4j EXE that knows how to open the Process Dashboard / Team
Dashboard, and adjust the windows shortcuts to point to that file
instead of javaw.