Use a Launch4J wrapper to uninstall the Process Dashboard
Recent changes by Oracle have caused shortcuts to javaw to fail on some
versions of Windows. To work around this problem, we now create a
Launch4j EXE that knows how to uninstall the Process Dashboard, and
adjust the windows shortcuts to point to that file instead of javaw.

In addition, the COIOSHelper DLL file (used by IzPack to create an entry
in the Add/Remove programs list) is now triggering a false positive
malware warning from Windows Defender. Such warnings are unacceptable,
so we have no choice but to turn off this IzPack feature. Users can
still uninstall the application with the appropriate shortcut icon.