WBS-wide workflow reapplies should require explicit opt-in
The Reapply Workflows operation can reapply workflow definitions across
the entire WBS. When choosing this option, it displays all of the
workflows in use and all of the enactments for each, with checkboxes to
indicate which items should be reapplied. Previously, these boxes were
all checked by default. This meant that an individual would need to
uncheck many items manually for the most common use case (fix a single
workflow). If they neglected to do so, clicking OK could have
far-reaching/unintended consequences. With this change, all items are
unchecked by default, and the user must specifically check a box to
request the reapplication of a particular item.