Display a PDES info link in the Team Project Setup Wizard
When a team is walking through the Team Project Setup Wizard and they
reach the page asking for the team data directory, they may find
themselves at a roadblock if they cannot create a shared network
directory. Display a note for these teams with a link to information
about the Process Dashboard Enterprise Server.