Keep local caches of network import directories
Bridged import directories (which come from the PDES) maintain a local
cache of imported files, and logic is able to use those cached files
when the network is unavailable. Historically, that has not been the
case for "regular" import directories (which are hosted on a network
filesystem). This change adds a local cache for those "regular"
directories, so non-PDES users can maintain access to imported files
even when they lose network connectivity.