Display better labels for the "before/after development" rows
Project plan summary forms in the dashboard have traditionally included
rows to show defects injected "before development" and removed "after
development." That was fine when the form was showing rows for the
entire development cycle, but it is confusing in the context of a
workflow report. ("Development" would generally be taken to describe all
steps in the lifecycle, not just the activities in this one workflow;
but we were giving it the latter meaning.) To rectify this problem, the
rows are now labelled "Before <first phase>" and "After <last phase>".