In bridged mode, handle the "import" subdir specially
By default, the dashboard imports files/data from an "import"
subdirectory that appears under the main data directory. In the past,
that subdirectory has never been present for a bridged dataset; but now,
the PDES may deliver files there. This change introduces a new
ImportDirectory implementation that knows how to refresh the contents of
that directory when needed.