Assign hierarchy node IDs immutably
Node IDs are used in the hierarchy by the PDES REST API. This change
arranges for:
 * The root node to have an ID of "0:root"
 * Team project roots to have node IDs of the form "abc123:root"
 * All other nodes to have plain numeric IDs.
 
Node IDs are no longer assigned eagerly; they must be explicitly enabled
via a call to a DashController method. This avoids introducing a
dependency on dashboard version 2.2.3.2 unless an individual is using
this functionality.