Retain WBS component size when all subtasks are deleted.
When a workflow is applied to a component, the component size estiate
may propagate down into a child task. Later, if those tasks are deleted,
the size estimate was being deleted along with the tasks. This was
problematic, so this change catches that scenario and retains the
estimated component size.