Stop timing a defect after active task change
If the defect timer is running, and the global active task changes,
check to see if the new active task "matches" the task where the defect
is logged. (Matching logic is configurable, but it defaults to whether
the tasks are part of the same component, workflow, or PSP project.) If
the tasks do not match, stop the defect timer.