Write data about new tasks in the userDump.xml file
The userDump.xml file (inside a PDASH export file) has historically
contained information about tasks that are locally added in the user's
plan. However, that information was not adequate to perform a reliable
reverse sync, so for years it has been ignored by the WBS Editor. This
change improves the export format, as a first step toward supporting
reverse sync of added tasks.