Bugfix for % Complete calc on PSP tasks with out-of-sync planned times
When an individual marks the planning phase complete on a PSP task in
their personal dashboard, the sync operation is no longer permitted to
copy down changes to the time estimate. This can result in time
estimates that are out of sync between the WBS and the personal
dashboard. When that scenario occurred, the % Complete column in the WBS
editor was being calculated incorrectly, sometimes resulting in PSP
tasks that were more than 100% complete.