Refactor the handling of "everyone" vs "all people" in group mgr
Several group filters and labels will display a concept of "Everyone" to
indicate that data from all users is being displayed. Since that is the
default mode of operation, we need ways to display and react to that
"group selection" quickly. Of course, "everyone" really means that no
group filter is being applied; so we don't need to retrieve the list of
all individuals to apply the "everyone" filter. This change follows that
observation and separates the logic to create an "everyone group" from
the time-consuming logic to "load all known individuals."