Include subschedule period data in saved EV baselines
When the dashboard saves a baseline for an EV schedule, it does so by
storing an XML representation of the given task list. In the past, those
XML representations have included detail about every task, but only
summarized information about the periods in the top-level schedule. This
change writes period data for each of the embedded subschedules.