In the PROBE Wizard, advise users that an Export may be needed
When run against a workflow, the PROBE Wizard uses data from PDASH files
that have been exported in the past. Display a footer on the historical
data page explaining this fact to users, and advising them to perform an
"Export My Data Now" operation if they feel that any data is missing.