On the Workflow Analysis, advise users that an Export may be needed
When run in the personal dashboard, the Workflow Process Analysis report
uses data from PDASH files that have been exported in the past. Display
a footer explaining this fact to users, and advising them to perform an
"Export My Data Now" operation if they feel that any data is missing.