Don't allow offline lock mode for Team Dashboards
Offline lock mode is a useful feature for individuals, but it doesn't
make much sense for Team Dashboards. (Keeping a Team Dashboard locked
for an extended period of time is bad for other users, and there aren't
many meaningful editing operations that a user could perform in a Team
Dashboard while it is offline.) This change hides the offline mode
checkbox for Team Dashboards, unless it is already locked; then, the
checkbox is displayed just until the user unlocks it.