Respect location remappings when republishing settings.xml files
If the user has opened a data backup with the Quick Launcher, make
certain the settings.xml republishing logic only touches the files that
came from the backup (not the real files out on the network). Similarly,
if they dragged an extracted directory onto the QuickLauncher, be
cognizant of directory remappings when republishing.