Disable charts on T&S window if data privacy permissions dictate
Users can be granted permission to view personal EV charts for specific
groups of individuals. If a group filter is applied in the Task &
Schedule window, and that filter would narrow the data down to a single
individual, disable the Charts button unless that individual is on the
permitted list.