Block group filters on the EV report if they violate data privacy
If the user applies a group filter to an EV report or chart gallery, and
that filter would narrow the view to a single individual, check to see
if the user has permission to see the EV report/charts for the
individual in question. If not, block the report to protect privacy.