Read known users from team data directories rather than the database
In the previous version of the dashboard, the User Group Manager
performed a database query to determine the list of all known people.
This change arranges for the list to be built from imported PDASH files,
and from the team.xml files that are present in the same directories.
This allows the list of people:
 * to be available immediately, without waiting for all projects to be
   loaded into the warehouse
 * to include individuals who have been added to a project, but who
   have not joined the project yet