Bugfix: WBS Editor must exit after bad password with data security off
If data security is disabled on the PDES, the first password prompt will
be triggered when the permissions manager connects to the "whoami" REST
API. If the user did not provide a valid password at that prompt, the
WBS Editor was not exiting as it should. This change fixes that bug.