Adjust WBS column calculations to respect the filter in effect
When a filter was applied to the WBS in the past, the only effect was
that nonmatching rows were hidden from display. With this change,
the hidden rows:
 * no longer contribute to bottom-up rollups for parent nodes
 * are no longer altered by top-down edits on parent nodes

The end result is that a filtered WBS behaves much like a WBS that
doesn't contain the filtered nodes at all. The resulting views are more
informative from a reporting perspective, because they display rollups
of the nodes that match the filter. And they are more usable from an
editing perspective, because your edits do not (generally) have any
effect on nodes that you cannot see.

As a downstream change, certain calculations in the TeamTimePanel (which
were previously assuming that row "0" of the WBS would provide
project-level totals) have been updated to draw their data from a more
reliable source.