Remove the now superceded "task status" WBS filtering option
The WBS traditionally had an option to filter the WBS by task status,
with separate options for not started, in progress, and complete. Those
options were not as intuitive as the new "Show Completed Tasks" toggle.
This change removes the older, more complex filter, allowing ths simpler
"Show Completed Tasks" toggle to provide the requisite functionality.