Provide complete package versioning info in JAR manifests
The Java package versioning specification describes a number of
attributes that can be placed into the manifest file to capture version
info for the packages in a JAR. Java 9 requires all of these attributes
to be provided, or else all are ignored. This change stores values for
all of the version attributes, as required.