Store WBS structural attributes in a field to improve performance
Tree operations on the WBS (e.g. to get the parent or children of a
node) are extremely common. Previously, the calculated values of these
operations were saved as attributes in the WBSNode's attributes map.
This change stores them in a dedicated field instead. This allows us to
skip the hashtable lookup for these common values, improving performance
of calculations by about 15%.