Bugfix: unable to assign multiple roles in some non-English locales
If the user was in a locale that uses a comma as a decimal separator
(for example, French), the UserEditor would not allow you to enter more
than one role per user. In addition, if role names contained non-alpha
characters, the autocompletion logic would not allow you to type them.