Improve the logic for creating the "Me" item in the group filter
The selector to filter by person/group contains a "Me" item. In the
past, if the current user had no representation in the team member list,
that "Me" item would still be created, and would act as a match-nothing
filter. This was confusing to users, especially when they expected
themselves to be present in the team member list.

In a related change, matches are allowed between the username of the
logged in user and the initials of a team member, if this team is using
a "username" initials policy.