Display notifications more rapidly after they are added
The notification window cannot be displayed when the user is interacting
with a menu, because it would cause their popup menus to collapse.  In
the past, this problem was avoided by only displaying notifications
immediately after the Process Dashboard window gained focus. Thus, if
the user was in another application and they switched back to the
dashboard, notifications would be displayed.

One unfortunate side effect was that notifications would never be
displayed upon startup, and would not be displayed if the dashboard
window was active when the notification arrived.  These notifications
would only be shown after the user toggled away to another app and then
came back to the dashboard.  In situations where a user is expecting and
waiting for a notification (such as a "join project" message), this
delay was problematic.

This change allows notifications to be displayed immediately if the
dashboard window is active and if no popup menus are open. If either of
those conditions is false, the notification will be displayed as soon as
the conditions become true.