Use consistent background color for main toolbar on Windows LAF
The Windows look and feel uses a white background for JMenuBar
components (presumably so the menu bar blends with the window
decorations). Unfortunately, this causes the "C" menu and the task
navigation selector to stand out from the rest of the main personal
dashboard toolbar (possibly because these components use JMenuBar in a
nonstandard way). This change arranges for these components to use the
same background color as other components, improving overall appearance.