Bugfix: the "No Group" option on Team Setup Wizard did not work
When a Team Dashboard user is creating a new Team Project, they must
select a name and "group" (i.e. hierarchy parent) for the new project.
If they browse for a group, one of the options is "No Group", which maps
to the hierarchy root "/". Unfortunately, if they choose that option,
the setup wizard would fail. This change corrects that problem.