Use a "hamburger" icon for the configuration menu
The configuration menu provides access to a set of infrequently-used
features. Historically, the menu was abbreviated with the letter "C" to
be compact. Unfortunately, the letter "C" is unintuitive, so users have
no idea what the letter stands for, or what they might happen if they
click it. This change adorns the menu with the "hamburger" icon, which
is an industry-standard indicator for a menu of less-frequently-used
functionality.