Don't bury "Save As" button on chart customization window
The customizable charts have a "Save As" feature that allows a team to
create multiple instantiations of a given chart, each with their own
settings. This is a very powerful feature; but unfortunately, it was
previously only visible in a drop-down menu, where very few people would
find it. This change moves it to a button of its own.