Add a simple JAX-RS-style dispatch servlet for REST endpoints
JAX-RS is an industry standard platform for implementing REST APIs. But
during a first attempt at building a REST API for the Process Dashboard,
it was discovered that a near-empty WAR file based on Jersey and Jackson
would add a full three seconds to the startup sequence of the app, even
though 95% of users will never invoke any of the functionality it
contains.

In the interest of performance, this simplified replacement allows
registration of various REST endpoints, and dispatches requests to those
endpoints based on path-match rules.