Don't base standard icon size on the configuration menu
On startup, the dashboard calculates an optimal size for application
icons, with the goal of making the buttons on the main toolbar the same
height as the menu bar for selecting the active task.

In the past, this was done by measuring the preferred height of the
configuration menu (since it is also a menu bar). But the config menu
now uses a hamburger icon rather than text, and this raises the
possibility that it could have a different preferred height on some
look-and-feels.

This change creates a dummy menu bar for performing the calculations, to
avoid than uncertainty that stems from the "hamburger" config menu.