Digitally sign JavaHelp add-on to allow trust by Java 9
In the past, the dashboard would build JavaHelp support by copying an
existing JAR file from the JavaHelp distribution, only adding a small
dashboard manifest to the META-INF directory. The JAR in the JavaHelp
distribution was signed by Sun Microsystems, and their signature was
retained unchanged. Their certificate was granted the necessary
permissions by the dashboard's security policy.

Unfortunately, the JavaHelp project has been abandoned, with no updates
since 2007; and the digital signature in that distribution uses an old
algorithm that is no longer trusted by Java 9. This causes the JavaHelp
window to be displayed with an "Applet Warning" indicator, and causes
the print buttons to fail.

To reestablish trust when running under Java 9, we must re-sign the
JavaHelp JAR file, rather than relying on Sun's old signature.