Allow dashboard web apps to respond to PUT requests
In REST APIs, it is a best practice to use PUT requests for idempotent
modifications to resources. Unfortunately, the prior logic in the
dashboard web server would receive a "method not allowed" error from the
first web app in the chain, and would not continue to try other apps.
This change resolves that problem, making it possible for web apps to
receive and respond to PUT requests.