New ability to generate less-cryptic redacted backups
Redaction is useful for generating backups that protect the privacy and
confidentiality of the teams that collected the data. But these backups
use hashcode-derived names for people, projects, and tasks, and those
names can be an assault on the brain if a person is trying to make sense
of the data in a demo. This change makes it possible to use
human-readable names for people and projects. It also improves the
generation of scrambled task list names, allowing them to mirror the
projects they describe.