Add a REST API to listen for async dashboard events
This change introduces the REST endpoint /api/v1/events/

Connections to this endpoint support the delivery of notifications about
events that have occurred in the dashboard, using a COMET-style long
polling mechanism. As part of this initial change, the endpoint
publishes notifications about changes to the state of the timer.