Make defect dialogs always-on-top, and emit JSON when opening via REST
Previously, defect dialogs used a JDialog, making them children of the
Process Dashboard toolbar. As a result, all open defect dialog windows
would disappear if the toolbar was minimized. In addition, if the
toolbar was hidden to the system tray, programmatically opening a defect
dialog would force the toolbar to reappear.

This change arranges to use always-on-top JFrames instead. The dialogs
appear in front of other windows (like IDEs), and do not disappear if
the Process Dashboard toolbar is minimized. They can also be shown by a
REST API call even when the dashboard toolbar is invisible.

Finally, when the /control/showDefectDialog endpoint is invoked, it will
emit a JSON reponse if one was requested in the Accept request header.