Add a control script that delegates to various trigger APIs
A number of dashboard scripts provide "trigger" functionality as a way
to invoke rich client actions from a URI endpoint. Although a client can
access any of these scripts directly, clients based on Retrofit or Refit
expect to have URI paths that are hardcoded in annotations. Those
libraries don't make it easy to invoke an arbitrary URI, selected at
runtime.

To work around that problem, this script accepts a trigger URI as a POST
body parameter, invokes that URI, and returns the response.