Refactor UserNotificationManager to use URIs instead of Runnables
In the original implementation of the UserNotificationManager,
notifications that required user action were registered with Runnable
objects, which would be executed when the user clicked a notification.
Unfortunately, that approach prevents handling of notifications by an
external UI (such as an IDE toolbar).

Fortunately, all existing notifications written to date are handled by
opening a URI in a web browser. This implementation replaces Runnable
actions with URIs. (These URIs can be trigger URIs, so rich client
responses are still possible.)