Cancel user notifications when the corresponding logic is run
When the user double-clicks on a notification in the dashboard UI, the
notification they clicked is removed from the list before its logic
runs. But if a notification is processed by an external client like an
IDE, it might not get removed from the notification list automatically.
This change updates the various notification handlers so they remove the
user notifications, just in case. This simplifies handling for the IDE,
which no longer needs to worry about deleting notifications as they are
handled.