Provide more usable labels and titles for defect-related UIs
In the past, the defect entry/editing window had the title "Defect
Dialog." That is a programmer-centric phrase, and is not user-friendly.
Now, the title tells the user whether they are adding or editing a
defect.

In the past, when an existing defect was edited, the path of the defect
was displayed in the title bar.  That was not user-friendly, because
most task names are too long to fit. Now, the defect dialog displays an
auto-abbreviating label showing the path the defect was logged to.

In the past, the Defect Log was called "Defect Editor".  That label was
inconsistent. Now, the window is labeled "Defect Log".