Remove the (virtually useless) MCF process script
The Team Project XML template includes a script-menu entry for a
"Process Script," which is nothing more than a bogus page with the
repeating text, "Customize this script...". Virtually no teams do that,
so the entry is dead weight.

This change removes the entry from the script menu. If a team needs the
entry to come back, it can be reenabled by constructing and installing
an empty dashboard package with a specific ID.